refactor: `s/manager/scheduler`
