+++ kernel/src/process/ipc.rs
@@ -1,1 +1,1 @@
-pub(crate) use super::manager::{receive_from, receive_from_any, send};
+pub(crate) use super::scheduler::{receive_from, receive_from_any, send};
